wait until python 2.7 for windows :-)

+++ trunk/maps.py
@@ -910,4 +910,5 @@ def main():
 if __name__ == "__main__":
     main()
     pid = os.getpid()
-    os.kill(pid, signal.SIGQUIT)
+    if os.name == 'posix':
+        os.kill(pid, signal.SIGQUIT)
